Smooth Overwrites
Added the Smooth Overwrites patch, with a way to distribute it this
time. I've put the patch in the resource pack zip, so that it can be
easily extracted into the resource pack folder and used. It's an extra
step for the users, and we need to get the word out that it's there and
how to use it, but it's the simplest way to do it.

Also tweaked a couple of ores.